changed "unsigned int" into "unsigned long" for s parameter
git-svn-id: svn://scm.gforge.inria.fr/svnroot/ecm/trunk@372 404564d9-a503-0410-82bf-e18ce2cf3989

+++ ecm.h
@@ -153,7 +153,7 @@ int          pm1         (mpz_t, mpz_t, mpz_t, double, double, double, double,
                           unsigned int, int, int, int);
 int     pm1_rootsF       (mpz_t, listz_t, unsigned int, mpres_t *, listz_t,
                           int, mpmod_t, int, unsigned long *);
-mpres_t *pm1_rootsG_init (mpres_t *, unsigned int, unsigned int, int,
+mpres_t *pm1_rootsG_init (mpres_t *, unsigned long, unsigned int, int,
                           mpmod_t);
 void    pm1_rootsG_clear (mpres_t *, int, mpmod_t);
 int     pm1_rootsG       (mpz_t, listz_t, unsigned int, mpres_t *, listz_t, 
@@ -175,7 +175,7 @@ int trial_factor(mpcandi_t *n, double maxfact, int deep);
 /* ecm2.c */
 int     ecm_rootsF       (mpz_t, listz_t, unsigned int, curve *,
                           int, mpmod_t, int, unsigned long *);
-point * ecm_rootsG_init  (mpz_t, curve *, unsigned int, unsigned int, 
+point * ecm_rootsG_init  (mpz_t, curve *, unsigned long, unsigned int,
                           int, mpmod_t, int);
 void    ecm_rootsG_clear (point *, int, mpmod_t);
 int     ecm_rootsG       (mpz_t, listz_t, unsigned int, point *,
@@ -190,7 +190,7 @@ int          pp1         (mpz_t, mpz_t, mpz_t, double, double, double, double,
                           unsigned int, unsigned int, int, int);
 int   pp1_rootsF         (listz_t, unsigned int, mpres_t *, listz_t,
                           mpmod_t, int, unsigned long *);
-mpres_t *pp1_rootsG_init (mpres_t *, unsigned int, unsigned int, mpmod_t);
+mpres_t *pp1_rootsG_init (mpres_t *, unsigned long, unsigned int, mpmod_t);
 void  pp1_rootsG_clear   (mpres_t *, mpmod_t);
 int   pp1_rootsG         (listz_t, unsigned int, mpres_t *, mpmod_t,
                           unsigned long *);
@@ -198,7 +198,7 @@ int   pp1_rootsG         (listz_t, unsigned int, mpres_t *, mpmod_t,
 /* stage2.c */
 int          stage2     (mpz_t, void *, mpmod_t, double, double, unsigned int, 
                          int, int, int);
-void  fin_diff_coeff    (listz_t coeffs, unsigned int s, unsigned int D, 
+void  fin_diff_coeff    (listz_t coeffs, unsigned long s, unsigned int D, 
                          unsigned int E, int dickson_a);
 
 /* listz.c */
